requirements.txt
Django>=4.2
django-widget-tweaks
google-generativeai
django-bootstrap5
